レビュー指摘反映：Adapterの処理を別ファイルに切り出す
